Make pFq inherit from CMF
