+++ ramanujantools/cmf/known_cmfs/__init__.py
@@ -20,8 +20,9 @@ from .known_cmfs import (
     cmf3_3,
     hypergeometric_derived_2F1,
     hypergeometric_derived_3F2,
-    pFq,
 )
+
+from .pfq import pFq
 
 __all__ = [
     "c0",
